Remove GRUB_DEFAULT dirty bit (not needed).
Even if the user doesn't use GRUB_DEFAULT=saved along with the
grub-set-default mechanism, always convert the configuration to use it.

It might be trickier for the user to figure out which entry is the
default through a text editor when GRUB_DEFAULT is set to 'saved', but
through the GUI it's absolutely clear.

Obvisouly, if the GUI is being used, the user won't have to resort to a
text editor.