Fix installation detection in Gentoo.
On Gentoo (probably elsewhere too), the GRUB2 executables are located
under /usr/sbin/ which is not included in the user's PATH by default.
The helper always has empty PATH so performing the detection there won't
help either. So use `whereis` as a fallback; not very nice a solution:
might consider something more elegant in the future.