*ADDED: Preliminary submenu support.
This is a first attempt at implementing submenu support. The parsing
code et al. is just fine, but the QComboBox+QTreeView combination leaves
a lot to be desired.

Specifically, I couldn't find a proper way to match the currently
selected item in the view with the corresponding QModelIndex. QComboBox
only advertises the current text and index, which is not enough for a
unique '1-1' mapping. The current QPersistentModelIndex is hidden in the
private implementation, no can do about it.

Perhaps a different widget combination could yield better results.