Oops. Multiple definitions in different translation units.
Inline the functions in order to share the definition in all translation
units.

Making these functions static is *wrong*! Each translation unit would
have a different copy of the function and thus different static local
variables. This kills the purpose of caching the value in a single
static local variable. Construct-On-First-Use would still work but with
different copies of the variable.

An excellent summary of the avaiable options can be found here:
http://stackoverflow.com/a/10847368/1266511